Melhora organização das pastas
